+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -3,6 +3,9 @@ package kitchenpos.application;
 import kitchenpos.domain.*;
 import kitchenpos.menu.application.InMemoryMenuRepository;
 import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.ordertable.application.InMemoryOrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
