+++ src/test/java/kitchenpos/ordertable/application/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.ordertable.application;
 
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 
 import java.util.*;
